added colors to updatejournal form
