+++ views/layouts/main.handlebars
@@ -14,7 +14,7 @@
     <link
         href="https://fonts.googleapis.com/css2?family=Karla:wght@300;400&family=Merriweather:wght@300;400&display=swap"
         rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="bulma/bulma.css">
+    <link rel="stylesheet" type="text/css" href="/bulma/bulma.css">
     <link rel="stylesheet" href="/css/style.css">
     <script src="https://kit.fontawesome.com/eb432245f2.js" crossorigin="anonymous"></script>
 </head>
@@ -23,7 +23,7 @@
     <header class="green" style="has-background-grey-lighter">
         <h1 class="title is-family-secondary is-size-2-mobile is-size-1 has-text-centered p-3">Bullet Journal</h1>
         {{#if logged_in}}
-        <nav class="navbar" role="navigation" aria-label="main navigation">
+        <nav class="navbar green" role="navigation" aria-label="main navigation">
             <div class="navbar-menu">
                 <div class="navbar-start">
                     <a href="/dashboard" class="navbar-item is-family-secondary">
