+++ views/partials/journal-partial.handlebars
@@ -1,8 +1,10 @@
 <div class=" column is-half-tablet is-one-third-desktop is-one-quarter-widescreen">
     <div class="card">
         <div class="{{background}} card-content">
+            <a href="/journals/{{id}}" >
             <p class="title"> Journal Title: {{title}}</p>
             <p class="subtitle">Journal Description {{description}}</p>
+            </a>
         </div>
     </div>
 </div>
